requirements.txt
aiohttp==3.7.4.post0
async-generator==1.10
async-timeout==3.0.1
attrs==21.4.0
certifi==2021.10.8
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
cryptography==36.0.1
dhooks==1.1.4
discord.py-self==1.9.1
emoji==1.6.3
h11==0.13.0
idna==3.3
loguru==0.6.0
multidict==6.0.2
oauthlib==3.2.0
outcome==1.1.0
pio==0.0.3
pycparser==2.21
pyOpenSSL==22.0.0
PyQt6==6.2.3
PyQt6-Qt6==6.2.3
PyQt6-sip==13.2.1
PySide6==6.2.3
PySocks==1.7.1
requests==2.27.1
requests-oauthlib==1.3.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
selenium==4.1.2
shiboken6==6.2.3
sniffio==1.2.0
sortedcontainers==2.4.0
trio==0.20.0
trio-websocket==0.9.2
tweepy==4.6.0
typing_extensions==4.1.1
urllib3==1.26.8
wsproto==1.0.0
yarl==1.7.2
